Advent of Code 2022: Day2
Similar to day1 there are two challs, the first chall is calculating our final score in a game of rock, papers and scissors. With a specific scoring system and a list of hands played. The second chall had the same requirement but instead of player hand each round, we have the outcome of that round. Basing on the outcome calculate the correspondant player move. the scoring function remains the same. Added [c_modulus] (calc_modul or crazy_modulus x)) taking to opponent hand and desired outcome and returns the players hand. feeding that return to the scoring function [s_game] to add the round's score to the final score.